Change deprecated numpy type

+++ environments/environment_abstract.py
@@ -17,7 +17,7 @@ class State(ABC):
 
 class Environment(ABC):
     def __init__(self):
-        self.dtype = np.float
+        self.dtype = float
         self.fixed_actions: bool = True
 
     @abstractmethod
